requirements.txt
langchain-groq
langchain-community
langchain-tavily
python-dotenv
uvicorn
fastapi
pydantic
streamlit
langgraph
langchain-core